Move mixins to global space (#6472)
