+++ packages/accessibility/global.d.ts
@@ -1,0 +1,8 @@
+declare namespace GlobalMixins
+{
+    // eslint-disable-next-line @typescript-eslint/no-empty-interface
+    interface DisplayObject extends Partial<import('@pixi/accessibility').IAccessibleTarget>
+    {
+
+    }
+}
